Fix: Item Detail Views not showing
This is due to the fact that we only have a RouterOutlet in our main AppView. Children of a router always use their
parent to render, as the parent doesn't have a RouterOutlet it cannot render the child and the parent stays visible.
This is additionally confusing because the routing is succeeding.